feat: create useAsyncThunk hook
