+++ ui/src/pages/organizations/index.tsx
@@ -1,17 +1,15 @@
 import classNames from 'classnames/bind';
 import Link from 'next/link';
-import { useEffect, useState, useCallback } from 'react';
 import { MdAdd, MdBusiness } from 'react-icons/md';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 import NoItemsFound from 'src/components/NoItemsFound';
 import OrganizationsList from 'src/components/Organization/List';
 import DashboardLayout from 'src/containers/DashboardLayout';
 import { fetchOrganizations, orgSelector } from 'src/duck/organizations';
+import { useAsyncThunk } from 'src/lib/useAsyncThunk';
 
 import style from './style.module.scss';
-import { unwrapResult } from '@reduxjs/toolkit';
-import { useFlash } from 'src/duck/flashMessages';
 
 const c = classNames.bind(style);
 
@@ -27,28 +25,8 @@ const Header: React.FC = () => (
 );
 
 const Index = () => {
-  const dispatch = useDispatch();
   const organizations = useSelector(orgSelector.selectAll);
-  const [flash] = useFlash();
-  const [isFetching, setIsFetching] = useState<boolean>(false);
-
-  const fetchOrgs = useCallback(async () => {
-    try {
-      setIsFetching(true);
-      await unwrapResult((await dispatch(fetchOrganizations())) as any);
-      setIsFetching(false);
-    } catch (err) {
-      flash({
-        body: err.message,
-        title: 'Failed to fetch some organizations',
-        type: 'error',
-      });
-    }
-  }, []);
-
-  useEffect(() => {
-    fetchOrgs();
-  }, [fetchOrgs]);
+  const [isFetching] = useAsyncThunk(fetchOrganizations, 'Failed to fetch some Organizations');
 
   if (!organizations.length && !isFetching) {
     return (
